refractor(Entities): refractored names and methods to meet minimum requirements for code standards
